+++ internal/gfm/gfm.go
@@ -1,4 +1,4 @@
-package common
+package gfm
 
 import (
 	"bytes"
