+++ internal/jsruntime/deno.go
@@ -1,4 +1,4 @@
-package common
+package jsruntime
 
 import (
 	"archive/zip"
