+++ server/build_analyzer.go
@@ -9,6 +9,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/esm-dev/esm.sh/internal/npm"
 	"github.com/ije/gox/set"
 )
 
@@ -20,17 +21,17 @@ type Ref struct {
 func (ctx *BuildContext) analyzeSplitting() (err error) {
 	exportNames := set.New[string]()
 
-	for _, exportName := range ctx.pkgJson.Exports.keys {
+	for _, exportName := range ctx.pkgJson.Exports.Keys() {
 		exportName := stripEntryModuleExt(exportName)
 		if (exportName == "." || (strings.HasPrefix(exportName, "./") && !strings.ContainsRune(exportName, '*'))) && !endsWith(exportName, ".json", ".css", ".wasm", ".d.ts", ".d.mts", ".d.cts") {
-			v := ctx.pkgJson.Exports.values[exportName]
+			v, _ := ctx.pkgJson.Exports.Get(exportName)
 			if s, ok := v.(string); ok {
 				if endsWith(s, ".json", ".css", ".wasm", ".d.ts", ".d.mts", ".d.cts") {
 					continue
 				}
-			} else if obj, ok := v.(JSONObject); ok {
+			} else if obj, ok := v.(npm.JSONObject); ok {
 				// ignore types only exports
-				if len(obj.keys) == 1 && obj.keys[0] == "types" {
+				if keys := obj.Keys(); len(keys) == 1 && keys[0] == "types" {
 					continue
 				}
 			}
